Add a ghpagesBranch setting.
This contains the name of the branch in which to put site data. It defaults to gh-pages, so that the behaviour of sbt-ghpages still defaults to proper githug pages support.

The purpose of this setting is to let people working with other systems than github (gitlab, in my case) use sbt-ghpages with very little additional configuration.